+++ engine/src/main/java/org/destinationsol/systems/DamageSystem.java
@@ -20,7 +20,7 @@ import org.destinationsol.components.Health;
 import org.destinationsol.entitysystem.EntitySystemManager;
 import org.destinationsol.entitysystem.EventReceiver;
 import org.destinationsol.events.DamageEvent;
-import org.destinationsol.events.ZeroHealthEvent;
+import org.destinationsol.removal.ZeroHealthEvent;
 import org.terasology.gestalt.entitysystem.entity.EntityRef;
 import org.terasology.gestalt.entitysystem.event.EventResult;
 import org.terasology.gestalt.entitysystem.event.ReceiveEvent;
